finish them all

+++ lib/models/request/overtime/update_approval_overtime_request.dart
@@ -12,7 +12,7 @@ class UpdateApprovalOvertimeRequest {
   });
 
   String? action;
-  int? actualTime;
+  double? actualTime;
   String? noteApproval;
 
   factory UpdateApprovalOvertimeRequest.fromRawJson(String str) =>
